error handling

+++ src/services/categories/categories.service.ts
@@ -25,6 +25,7 @@ export const createCategory = async (data: categoryDTO) => {
             await cacheService.updateSingleCategoryCache(doc.queryResponse);
             // category list cache update
             await cacheService.updateCategoryListCache();
+            await cacheService.updateActiveCategoryListCache();
         }
         return doc;
     }
@@ -45,72 +46,97 @@ export const createCategory = async (data: categoryDTO) => {
 // Get single category
 export const getCategory = async (req: express.Request, res: express.Response) => {
     const id = req.params.id
+    //From db
+    console.log("from bd")
     const category = await categoryHelper.getSingleCategory(id)
     console.log({category})
-    if (category) {
-        const response: IResult = {
-            success: true,
-            statusCode: 200,
-            queryResponse: category,
-            message: '',
-        };
-        return response
-    } else {
-        const response: IResult = {
-            success: false,
-            statusCode: 422,
-            queryResponse: null,
-            message: 'Invalid Category id',
-        };
-        return response
+    if (category.success===false) {
+        return category
+    }else{
+    // single category cache update
+    const response: IResult = {
+        success: true,
+        statusCode: 200,
+        queryResponse: category,
+        message: '',
+    };
+    
+    await cacheService.updateSingleCategoryCache(category);
+    return response
+
     }
+
+
 }
 // Get all categories
-export const allCategories = async (res:express.Response) => {
-    const DEFAULT_EXPIRATION = config.defaultCashExpiration
+export const allCategories = async (res: express.Response) => {
     const resp = await findAll({ model: Category, obj: {}, projection: { name: 1, isRoot: 1, leaf: 1, active: 1 } })
-    await redisClient.setEx('all_categories', DEFAULT_EXPIRATION, JSON.stringify(resp));
+    console.log("from bd")
+    // category list cache update
+    await cacheService.updateCategoryListCache();
     res.send(resp);
 }
 // Get all active categories only
 export const getAllActiveCategories = async () => {
+    console.log("from bd")
     const res = await findAll({ model: Category, obj: { active: true } })
+    // category list cache update
+    await cacheService.updateActiveCategoryListCache();
+    return res
+
+}
+// Get all deactive categories only
+export const getAllDeactiveCategories = async () => {
+    console.log("from bd")
+    const res = await findAll({ model: Category, obj: { active: false } })
+    // category list cache update
+    await cacheService.updateDeactiveCategoryListCache();
     return res
 
 }
 // Search category with all nested sub categories
-export const searchCategory = async (key: string, res:express.Response) => {
-    // //check in cache
-    // const cacheDate = await redisClient.get(key)
-    // if (cacheDate) return JSON.parse(cacheDate)
-
+export const searchCategory = async (key: string, res: express.Response) => {
+    //check in cache
+    console.log("from cached")
+    const cacheDate = await redisClient.get(key)
+    if (cacheDate) return JSON.parse(cacheDate)
+    console.log("from bd")
     const DEFAULT_EXPIRATION = config.defaultCashExpiration
     const category = await categoryHelper.categorySearch(key)
+    if (category.success === false) {
+        const response: IResult = {
+            success: false,
+            statusCode: 422,
+            queryResponse: [],
+            message: 'No Category Found',
+        };
+        return response
+    }
     const childCategory = await categoryHelper.getChildCategories(category)
     let doc;
     if (childCategory) {
         doc = { category, childCategory }
-        await redisClient.setEx(key, DEFAULT_EXPIRATION, JSON.stringify(doc));
         const response: IResult = {
             success: true,
             statusCode: 200,
             queryResponse: doc,
             message: '',
         };
+        await redisClient.setEx(key, DEFAULT_EXPIRATION, JSON.stringify(response));
         return response
     }
     doc = { category }
-    await redisClient.setEx(key, DEFAULT_EXPIRATION, JSON.stringify(doc));
     const response: IResult = {
         success: true,
         statusCode: 200,
         queryResponse: doc,
         message: '',
     };
+    await redisClient.setEx(key, DEFAULT_EXPIRATION, JSON.stringify(doc));
     return response
 }
 // Update Category
-export const updateCategory = async (req:express.Request, res:express.Response) => {
+export const updateCategory = async (req: express.Request, res: express.Response) => {
     const id = req.params.id
     const body = req.body
     const doc = await Category.updateOne({ _id: id }, body)
@@ -123,12 +149,27 @@ export const updateCategory = async (req:express.Request, res:express.Response) 
         await cacheService.updateSingleCategoryCache(singleDoc);
         // category list cache update
         await cacheService.updateCategoryListCache();
+        await cacheService.updateActiveCategoryListCache();
+        await cacheService.updateDeactiveCategoryListCache();
     }
     res.json(doc)
 }
 // Delete Category
 export const deleteCategory = async (req: express.Request) => {
     const id = req.params.id;
+    console.log(id)
+    const category = await categoryHelper.getSingleCategory(id)
+    console.log({category})
+    if (category.success===false) {
+        console.log(1)
+        const response: IResult = {
+            success: false,
+            statusCode: 422,
+            queryResponse: [],
+            message: 'In valid Category ID',
+        };
+        return response
+    }
     const doc = await Category.deleteOne({ _id: id })
     console.log({ doc })
     if (doc) {
@@ -136,6 +177,8 @@ export const deleteCategory = async (req: express.Request) => {
         await cacheService.deleteSingleCategoryFromCache(id);
         // category list cache update
         await cacheService.updateCategoryListCache();
+        await cacheService.updateActiveCategoryListCache();
+        await cacheService.updateDeactiveCategoryListCache();
         const response: IResult = {
             success: true,
             statusCode: 200,
@@ -153,14 +196,15 @@ export const deleteCategory = async (req: express.Request) => {
     return response
 }
 // Deactive parent and child_category associated with that parent
-export const deactiveParentCategories = async (req:express.Request, res:express.Response) => {
+export const deactiveParentCategories = async (req: express.Request, res: express.Response) => {
     const id = req.params.id
-    const action:boolean = false
-    await categoryHelper.activeOrDeactiveParentCategory(id, action, res);
+    const action: boolean = false
+    return await categoryHelper.activeOrDeactiveParentCategory(id, action, res);
+
 }
 // Active parent and child_category associated with that parent
-export const activeParentCategories = async (req:express.Request, res:express.Response) => {
+export const activeParentCategories = async (req: express.Request, res: express.Response) => {
     const id = req.params.id
-    const action:boolean = true
-    await categoryHelper.activeOrDeactiveParentCategory(id, action, res);
+    const action: boolean = true
+    return await categoryHelper.activeOrDeactiveParentCategory(id, action, res);
 }
